+++ src/app/(landing)/blog/page.tsx
@@ -1,0 +1,4 @@
+export default function Blog()
+{
+    return <h1>Blog Page</h1>
+}
