+++ src/app/(landing)/contact-us/page.tsx
@@ -1,0 +1,4 @@
+export default function ContactUs()
+{
+    return <h1>Contact US page</h1>
+}
