+++ src/app/(landing)/help/page.tsx
@@ -1,0 +1,4 @@
+export default function Help()
+{
+    return <h1>Help page</h1>
+}
