+++ src/app/(landing)/terms-condition/page.tsx
@@ -1,0 +1,4 @@
+export default function TermsAndConditions()
+{
+    return <h1>Terms & Conditions page</h1>
+}
